requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.0.1
python-multipart==0.0.6
aiofiles==0.7.0
aiomysql==0.1.1
python-dotenv==1.0.0
mysql-connector-python==8.0.26
aiohttp==3.9.1
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
requests==2.31.0
dnspython==2.4.2
cryptography==3.4.7
python-dateutil==2.8.2